Rational 2-factor: re-factor swaption pricing to unify forthcoming pricers (cap/floor)
